Moving to a microservice architecture
removed old ifcpipeline
