+++ ifc2json/requirements.txt
@@ -1,0 +1,4 @@
+fastapi
+pydantic
+uvicorn
+ifcopenshell
