+++ ifcconvert/requirements.txt
@@ -1,0 +1,3 @@
+fastapi
+pydantic
+uvicorn
