+++ ifccsv/requirements.txt
@@ -1,0 +1,5 @@
+fastapi
+pydantic
+uvicorn
+ifccsv
+ifcopenshell
