--- ifcpipeline/requirements.txt
@@ -1,11 +1,0 @@
-ifcopenshell
-ifctester
-ifccsv
-ifcclash
-ifcdiff
-uvicorn
-fastapi
-python-multipart
-requests
-pydantic
-scikit-learn
