+++ ifctester/requirements.txt
@@ -1,0 +1,6 @@
+fastapi
+pydantic
+uvicorn
+ifcopenshell
+ifctester
+pystache
